chore: fix clippy warnings and version tests for crates.io publish
- Fix redundant_closure in serdes-ai-mcp
- Fix too_many_arguments with #[allow] in openai stream
- Fix manual_strip with strip_prefix in anthropic stream
- Fix unnecessary_cast in mistral, ollama, huggingface
- Fix question_mark lint in anthropic stream
- Fix field_reassign_with_default in google, groq, mistral, ollama providers
- Fix collapsible_match in bedrock, claude_code_oauth
- Fix filter_map -> map in openrouter
- Fix identical if blocks in cohere, ollama
- Fix and_then -> map in chatgpt_oauth
- Fix vec_init_then_push in history tests
- Update version tests to match 0.1.2

+++ serdes-ai-models/src/huggingface/model.rs
@@ -172,7 +172,7 @@ impl HuggingFaceModel {
             params.stop = Some(stop.clone());
         }
         if let Some(seed) = settings.seed {
-            params.seed = Some(seed as u64);
+            params.seed = Some(seed);
         }
 
         // Don't return the prompt in the output
